hackgtcedar/requirements.txt
# Core Flask dependencies
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6

# Twilio integration
twilio>=9.0.0

# Environment and HTTP
python-dotenv==1.0.0
requests==2.31.0

# WebSocket support
python-socketio==5.8.0
eventlet==0.33.3
